split out the PID logic into a new pid lib
